convert tools to TypeScript
